+++ tests/prefetch/test_prefetch_anchor_pivot.py
@@ -50,7 +50,8 @@ async def create_test_database():
         yield
 
 
-async def test_prefetch_pivot_all():
+@pytest.mark.parametrize("target_prefix", ["", "+", "post__"])
+async def test_prefetch_pivot_all(target_prefix):
     user = await User.query.create(name="Edgy")
     post = await user.posts.create(body="edgy is the best ORM ever")
     await post.comments.create(body="is true")
@@ -67,18 +68,35 @@ async def test_prefetch_pivot_all():
     posts = (
         await Comment.query.select_related("post")
         .prefetch_related(
-            Prefetch(to_attr="comments_filtered", related_name="comments", from_anchor="post"),
-            Prefetch(to_attr="reactions_filtered", related_name="reactions", from_anchor="post"),
-            Prefetch(to_attr="users_filtered", related_name="user", from_anchor="post"),
+            Prefetch(
+                to_attr=f"{target_prefix}comments_filtered",
+                related_name="comments",
+                from_anchor="post",
+            ),
+            Prefetch(
+                to_attr=f"{target_prefix}reactions_filtered",
+                related_name="reactions",
+                from_anchor="post",
+            ),
+            Prefetch(
+                to_attr=f"{target_prefix}users_filtered", related_name="user", from_anchor="post"
+            ),
         )
         .update_embed_parent(("post", "origin_comment"))
     )
     assert len(posts) == 4
-    assert len(posts[0].users_filtered) == 1
-    assert len(posts[0].comments_filtered) == 2
-    assert len(posts[2].comments_filtered) == 2
-    assert len(posts[0].reactions_filtered) == 2
-    assert len(posts[2].reactions_filtered) == 2
+    if target_prefix:
+        assert len(posts[0].users_filtered) == 1
+        assert len(posts[0].comments_filtered) == 2
+        assert len(posts[2].comments_filtered) == 2
+        assert len(posts[0].reactions_filtered) == 2
+        assert len(posts[2].reactions_filtered) == 2
+    else:
+        assert len(posts[0].origin_comment.users_filtered) == 1
+        assert len(posts[0].origin_comment.comments_filtered) == 2
+        assert len(posts[2].origin_comment.comments_filtered) == 2
+        assert len(posts[0].origin_comment.reactions_filtered) == 2
+        assert len(posts[2].origin_comment.reactions_filtered) == 2
 
 
 async def test_prefetch_mixed_pivot1():
@@ -98,8 +116,8 @@ async def test_prefetch_mixed_pivot1():
         await Comment.query.select_related("post")
         .prefetch_related(
             Prefetch(to_attr="post__comments_filtered", related_name="post__comments"),
-            Prefetch(to_attr="reactions_filtered", related_name="reactions", from_anchor="post"),
-            Prefetch(to_attr="users_filtered", related_name="user", from_anchor="post"),
+            Prefetch(to_attr="+reactions_filtered", related_name="reactions", from_anchor="post"),
+            Prefetch(to_attr="+users_filtered", related_name="user", from_anchor="post"),
         )
         .update_embed_parent(("post", "origin_comment"))
     )
@@ -127,7 +145,9 @@ async def test_prefetch_mixed_pivot2():
     posts = (
         await Comment.query.select_related("post")
         .prefetch_related(
-            Prefetch(to_attr="comments_filtered", related_name="comments", from_anchor="post"),
+            Prefetch(
+                to_attr="post__comments_filtered", related_name="comments", from_anchor="post"
+            ),
             Prefetch(to_attr="post__reactions_filtered", related_name="post__reactions"),
             Prefetch(to_attr="post__users_filtered", related_name="post__user"),
         )
@@ -139,30 +159,3 @@ async def test_prefetch_mixed_pivot2():
     assert len(posts[2].comments_filtered) == 2
     assert len(posts[0].reactions_filtered) == 2
     assert len(posts[2].reactions_filtered) == 2
-
-
-async def test_prefetch_select_related_pivot():
-    user = await User.query.create(name="Edgy")
-    post = await user.posts.create(body="edgy is the best ORM ever")
-    await post.comments.create(body="is true")
-    await post.comments.create(body="I like javascript edgy")
-    await post.comments.create(body="I like python edgy")
-    await post.reactions.create(body="is true")
-    await post.reactions.create(body="I like javascript")
-
-    post = await user.posts.create(body="ravyn can handle asgi on a top level")
-    await post.comments.create(body="is true")
-    await post.comments.create(body="is really true")
-    await post.comments.create(body="totally true")
-    await post.reactions.create(body="is true")
-    await post.reactions.create(body="I like javascript")
-
-    posts = await Post.query.prefetch_related(
-        Prefetch(to_attr="comments_filtered", related_name="comments"),
-        Prefetch(to_attr="reactions_filtered", related_name="reactions"),
-    )
-    assert len(posts) == 2
-    assert len(posts[0].comments_filtered) == 3
-    assert len(posts[1].comments_filtered) == 3
-    assert len(posts[0].reactions_filtered) == 2
-    assert len(posts[1].reactions_filtered) == 2